Mobile h1h2css changes. Renaming components and path for further feature development.
