+++ src/app/modules/site/landing/highlight-block/highlight-block.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { HttpService } from '../../../commons/services/http/http.service';
+import { HttpService } from '../../../../commons/services/http/http.service';
 
 @Component({
   selector: 'app-highlight-block',
