+++ src/app/modules/site/landing/highlight-slider/highlight-slider.component.spec.ts
