+++ src/app/modules/site/landing/highlight-slider/highlight-slider.component.ts
@@ -1,5 +1,5 @@
 import { Component, AfterViewInit } from '@angular/core';
-import { HttpService } from '../../../commons/services/http/http.service';
+import { HttpService } from '../../../../commons/services/http/http.service';
 
 @Component({
   selector: 'app-highlight-slider',
