+++ src/app/modules/site/menubar/menubar.component.html
@@ -101,7 +101,7 @@
       </form>
     </div>
   </div>
-  <app-rendermd></app-rendermd>
+  <app-renderer></app-renderer>
 
   <span *ngIf="searchicon" class="fab-icon" title="Search Site" (click)="showSearch()">
     <img src="/assets/img/search.svg" class="searchicon-mob-img" alt="Search">
